Double Update

+++ Calculator/Main.c
@@ -1,29 +1,31 @@
 #include <stdio.h>
 
-int main() {
-    int a, b;
+int main()
+{
+    double a, b;
     char op;
     printf("Enter the first number: ");
-    scanf("%d", &a);
+    scanf("%lf", &a);
     printf("Enter the operator: ");
     scanf(" %c", &op);
     printf("Enter the second number: ");
-    scanf("%d", &b);
-    switch(op) {
-        case '+':
-            printf("The sum is %d\n", a + b);
-            break;
-        case '-':
-            printf("The difference is %d\n", a - b);
-            break;
-        case '*':
-            printf("The product is %d\n", a * b);
-            break;
-        case '/':
-            printf("The quotient is %d\n", a / b);
-            break;
-        default:
-            printf("Invalid operator\n");
+    scanf("%lf", &b);
+    switch (op)
+    {
+    case '+':
+        printf("The sum is %lf\n", a + b);
+        break;
+    case '-':
+        printf("The difference is %lf\n", a - b);
+        break;
+    case '*':
+        printf("The product is %lf\n", a * b);
+        break;
+    case '/':
+        printf("The quotient is %lf\n", a / b);
+        break;
+    default:
+        printf("Invalid operator\n");
     }
     return 0;
 }
